% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{load_app}
\alias{load_app}
\title{Load OpenCPU application}
\usage{
load_app(repo, ...)
}
\arguments{
\item{repo}{either name of a locally installed package, or a repository
address (passed to \code{install_github})}

\item{...}{passed to \code{ocpu_start}}
}
\description{
Loads an OpenCPU app from your local library or a github repository.
}
