% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{ocpu_start}
\alias{ocpu_start}
\alias{opencpu}
\alias{ocpu}
\title{OpenCPU Single-User Server}
\usage{
ocpu_start(port = 9999, root = "/ocpu", workers = 2, preload = NULL)
}
\arguments{
\item{port}{port number}

\item{root}{base of the URL where to host the OpenCPU API}

\item{workers}{number of worker processes}

\item{preload}{character vector of packages to preload in the workers. This speeds
up requests to those packages.}
}
\description{
Starts the OpenCPU single-user server for developing and testing apps locally.
}
