% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpd.R
\name{start_server}
\alias{start_server}
\alias{opencpu}
\alias{ocpu}
\title{OpenCPU Single-User Server}
\usage{
start_server(port = 9999, root = "/ocpu", workers = 2, preload = NULL,
  on_startup = NULL)
}
\arguments{
\item{port}{port number}

\item{root}{base of the URL where to host the OpenCPU API}

\item{workers}{number of worker processes}

\item{preload}{character vector of packages to preload in the workers. This speeds
up requests to those packages.}

\item{on_startup}{function to call once server has started (e.g. \code{browseURL})}
}
\description{
Starts the OpenCPU single-user server for developing and testing apps locally.
}
