% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{download_apps}
\alias{download_apps}
\alias{apps}
\alias{remove_apps}
\alias{app_info}
\alias{installed_apps}
\title{OpenCPU Application}
\usage{
download_apps(repo, ...)

remove_apps(repo)

app_info(repo)

installed_apps()
}
\arguments{
\item{repo}{a github repository such as \code{user/repo}, see \link{install_github}.}

\item{...}{additional options for \code{install_github}}
}
\description{
Manage locally installed OpenCPU applications. These applications can be
started via \link{start_github_app}.
}
\details{
These functions are used only for hosting locally using the single-user server.
To install apps on cloud servers and \code{ocpu.io} we use Github webhooks.
}
