% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{apps}
\alias{apps}
\alias{install_apps}
\alias{remove_apps}
\alias{installed_apps}
\alias{available_apps}
\title{OpenCPU Application}
\usage{
install_apps(repo, ...)

remove_apps(repo)

installed_apps()

available_apps()
}
\arguments{
\item{repo}{a github repository such as \code{user/repo}, see \link{install_github}.}

\item{...}{additional options for \code{install_github}}
}
\description{
Manage installed OpenCPU applications. These applications can be started locally
using \link{ocpu_start_app}.
}
\details{
When running this functions as \code{root} user, they can be used to manage globally
installed apps on an OpenCPU cloud server. However to install apps on cloud servers and
\code{ocpu.io} it is easier to the Github webhook.
}
\examples{
\dontrun{
# List available demo apps
available_apps()

# Run application from: https://github.com/rwebapps/stocks
ocpu_start_app("rwebapps/stocks")

# Run application from: https://github.com/rwebapps/markdownapp
ocpu_start_app("rwebapps/markdownapp")

# Run application from: https://github.com/rwebapps/appdemo
ocpu_start_app("rwebapps/appdemo")

# Show currently installed apps
installed_apps()
}
}
\seealso{
Other ocpu: \code{\link{ocpu-server}}
}
