% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{start_github_app}
\alias{start_github_app}
\alias{apps}
\alias{start_local_app}
\title{Load OpenCPU application}
\usage{
start_github_app(repo, ...)

start_local_app(package, ...)
}
\arguments{
\item{repo}{either name of a locally installed package, or a repository
address (passed to \code{install_github})}

\item{...}{passed to \code{start_server}}

\item{package}{name of locally installed package}
}
\description{
Loads an OpenCPU app from your local library or a github repository.
}
