% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{download_apps}
\alias{download_apps}
\alias{apps}
\alias{remove_apps}
\alias{info_apps}
\alias{installed_apps}
\title{OpenCPU Application}
\usage{
download_apps(repo, ...)

remove_apps(repo)

info_apps(repo)

installed_apps()
}
\arguments{
\item{repo}{a github repository such as \code{user/repo}, see \link{install_github}.}

\item{...}{additional options for \code{install_github}}
}
\description{
Manage installed OpenCPU applications. These applications can be started locally
using \link{start_github_app}.
}
\details{
When running this functions as \code{root} user, they can be used to manage globally
installed apps on an OpenCPU cloud server. However to install apps on cloud servers and
\code{ocpu.io} it is easier to the Github webhook.
}
