% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{apps}
\alias{apps}
\alias{ocpu_install_apps}
\alias{ocpu_remove_apps}
\alias{ocpu_installed_apps}
\alias{ocpu_available_apps}
\title{OpenCPU Application}
\usage{
ocpu_install_apps(repo, ...)

ocpu_remove_apps(repo)

ocpu_installed_apps()

ocpu_available_apps()
}
\arguments{
\item{repo}{a github repository such as \code{user/repo}, see \link{install_github}.}

\item{...}{additional options for \code{install_github}}
}
\description{
Manage installed OpenCPU applications. These applications can be started locally
using \link{ocpu_start_app}.
}
\details{
When running this functions as \code{root} user, they can be used to manage globally
installed apps on an OpenCPU cloud server. However to install apps on cloud servers and
\code{ocpu.io} it is easier to the Github webhook.
}
\seealso{
Other ocpu: \code{\link{ocpu-server}}
}
